move conversion types to own file, add types
