+++ src/state/trajectory/selectors/basic.ts
@@ -14,3 +14,4 @@ export const getSimulariumFile = (state: State) =>
     state.trajectory.simulariumFile;
 export const getAgentDisplayNamesAndStates = (state: State) =>
     state.trajectory.agentUiNames;
+export const getConversionProcessingData = (state: State) => state.trajectory.processingData;
